+++ packages/core/src/cli/validate/upgradeability-assessment.ts
@@ -1,25 +1,8 @@
 import { getAnnotationArgs, getDocumentation } from '../../utils/annotations';
 import { inferInitializable, inferUUPS } from '../../validate/query';
 import { ValidateCommandError } from './error';
+import { findContract } from './find-contract';
 import { SourceContract } from './validations';
-
-export class ReferenceContractNotFound extends Error {
-  /**
-   * The contract reference that could not be found.
-   */
-  readonly reference: string;
-
-  /**
-   * The fully qualified name of the contract that referenced the missing contract.
-   */
-  readonly origin: string;
-
-  constructor(reference: string, origin: string) {
-    super(`Could not find contract ${reference} referenced in ${origin}.`);
-    this.reference = reference;
-    this.origin = origin;
-  }
-}
 
 interface AnnotationAssessment {
   upgradeable: boolean;
@@ -35,6 +18,7 @@ export interface UpgradeabilityAssessment {
 export function getUpgradeabilityAssessment(
   contract: SourceContract,
   allContracts: SourceContract[],
+  overrideReferenceContract?: SourceContract,
 ): UpgradeabilityAssessment {
   const fullContractName = contract.fullyQualifiedName;
   const contractValidation = contract.validationData[fullContractName];
@@ -42,42 +26,26 @@ export function getUpgradeabilityAssessment(
   const isUUPS = inferUUPS(contract.validationData, fullContractName);
 
   const annotationAssessment = getAnnotationAssessment(contract);
-  if (annotationAssessment.upgradeable) {
-    let referenceContract = undefined;
-    let isReferenceUUPS = false;
-    if (annotationAssessment.referenceName !== undefined) {
-      referenceContract = getReferenceContract(annotationAssessment.referenceName, contract, allContracts);
-      isReferenceUUPS = inferUUPS(referenceContract.validationData, referenceContract.fullyQualifiedName);
-    }
 
-    return {
-      upgradeable: true,
-      referenceContract: referenceContract,
-      uups: isReferenceUUPS || isUUPS,
-    };
-  } else {
-    const initializable = inferInitializable(contractValidation);
-    return {
-      upgradeable: initializable || isUUPS,
-      uups: isUUPS,
-    };
+  let referenceContract = overrideReferenceContract;
+  if (referenceContract === undefined && annotationAssessment.referenceName !== undefined) {
+    referenceContract = findContract(annotationAssessment.referenceName, contract, allContracts);
   }
-}
 
-function getReferenceContract(reference: string, origin: SourceContract, allContracts: SourceContract[]) {
-  const referenceContracts = allContracts.filter(c => c.fullyQualifiedName === reference || c.name === reference);
-
-  if (referenceContracts.length > 1) {
-    throw new ValidateCommandError(
-      `Found multiple contracts with name ${reference} referenced in ${origin.fullyQualifiedName}.`,
-      () =>
-        `This may be caused by old copies of build info files. Clean and recompile your project, then run the command again with the updated files.`,
-    );
-  } else if (referenceContracts.length === 1) {
-    return referenceContracts[0];
-  } else {
-    throw new ReferenceContractNotFound(reference, origin.fullyQualifiedName);
+  let isReferenceUUPS = false;
+  if (referenceContract !== undefined) {
+    isReferenceUUPS = inferUUPS(referenceContract.validationData, referenceContract.fullyQualifiedName);
   }
+
+  return {
+    upgradeable:
+      referenceContract !== undefined ||
+      annotationAssessment.upgradeable ||
+      inferInitializable(contractValidation) ||
+      isUUPS,
+    referenceContract: referenceContract,
+    uups: isReferenceUUPS || isUUPS,
+  };
 }
 
 function getAnnotationAssessment(contract: SourceContract): AnnotationAssessment {